backend/requirements.txt
fastapi
uvicorn
python-multipart
sqlalchemy
groq
python-dotenv
pymupdf
python-docx
reportlab
passlib[bcrypt]
python-jose[cryptography]
pydantic[email]
fastapi-mail
